feat(project): improve settings
