--- apps/frontend/src/containers/Project/Domain.tsx
@@ -1,465 +1,0 @@
-import { useApolloClient } from "@apollo/client/react";
-import { SLUG_REGEX } from "@argos/util/slug";
-import { DialogTrigger } from "react-aria-components";
-import { SubmitHandler, useForm } from "react-hook-form";
-
-import { config } from "@/config";
-import { DocumentType, graphql } from "@/gql";
-import { DeploymentAuth } from "@/gql/graphql";
-import { Button } from "@/ui/Button";
-import {
-  Card,
-  CardBody,
-  CardFooter,
-  CardParagraph,
-  CardTitle,
-} from "@/ui/Card";
-import { Code } from "@/ui/Code";
-import {
-  Dialog,
-  DialogBody,
-  DialogDismiss,
-  DialogFooter,
-  DialogText,
-  DialogTitle,
-  useOverlayTriggerState,
-} from "@/ui/Dialog";
-import { FieldError } from "@/ui/FieldError";
-import { Form } from "@/ui/Form";
-import { FormCardFooter } from "@/ui/FormCardFooter";
-import { FormSwitch } from "@/ui/FormSwitch";
-import { FormTextInput } from "@/ui/FormTextInput";
-import {
-  ListBox,
-  ListBoxItem,
-  ListBoxItemDescription,
-  ListBoxItemLabel,
-} from "@/ui/ListBox";
-import { Modal } from "@/ui/Modal";
-import { Popover } from "@/ui/Popover";
-import { SelectButton, SelectField } from "@/ui/Select";
-
-import { getRepositoryLabel } from "../Repository";
-
-const INTERNAL_DOMAIN_SUFFIX = config.deployments.baseDomain;
-type DeploymentAuthLevel =
-  | DeploymentAuth.DomainPrivate
-  | DeploymentAuth.Private;
-
-const _ProjectFragment = graphql(`
-  fragment ProjectDomain_Project on Project {
-    id
-    domain
-    deploymentEnabled
-    deploymentAuth
-    customDeploymentProductionBranchGlob
-    repository {
-      __typename
-      id
-      defaultBranch
-    }
-  }
-`);
-
-const UpdateProjectDomainMutation = graphql(`
-  mutation ProjectDomain_updateProjectDomain(
-    $input: UpdateProjectDomainInput!
-  ) {
-    updateProjectDomain(input: $input) {
-      id
-      domain
-    }
-  }
-`);
-
-const UpdateProjectMutation = graphql(`
-  mutation ProjectDomain_updateProject(
-    $projectId: ID!
-    $deploymentProductionBranchGlob: String
-    $deploymentEnabled: Boolean
-    $deploymentAuth: DeploymentAuth
-  ) {
-    updateProject(
-      input: {
-        id: $projectId
-        deploymentProductionBranchGlob: $deploymentProductionBranchGlob
-        deploymentEnabled: $deploymentEnabled
-        deploymentAuth: $deploymentAuth
-      }
-    ) {
-      id
-      customDeploymentProductionBranchGlob
-      deploymentEnabled
-      deploymentAuth
-    }
-  }
-`);
-
-export function ProjectDomain(props: {
-  project: DocumentType<typeof _ProjectFragment>;
-  isTeam: boolean;
-}) {
-  const { project, isTeam } = props;
-
-  if (project.deploymentEnabled) {
-    return <ProjectDeploymentsEnabled project={project} isTeam={isTeam} />;
-  }
-
-  return <ProjectDeploymentsDisabled project={project} />;
-}
-
-type Inputs = {
-  domain: string;
-  noCustomDeploymentProductionBranchGlob: boolean;
-  deploymentProductionBranchGlob: string;
-  deploymentAuthEnabled: boolean;
-  deploymentAuth: DeploymentAuthLevel;
-};
-
-function ProjectDeploymentsEnabled(props: {
-  project: DocumentType<typeof _ProjectFragment>;
-  isTeam: boolean;
-}) {
-  const { project, isTeam } = props;
-  const client = useApolloClient();
-  const defaultDeploymentProductionBranchGlob =
-    project.repository?.defaultBranch || "main";
-  const form = useForm<Inputs>({
-    defaultValues: {
-      domain: getDomainSlug(project.domain),
-      noCustomDeploymentProductionBranchGlob:
-        project.customDeploymentProductionBranchGlob === null,
-      deploymentProductionBranchGlob:
-        project.customDeploymentProductionBranchGlob ||
-        defaultDeploymentProductionBranchGlob,
-      deploymentAuthEnabled: project.deploymentAuth !== DeploymentAuth.Public,
-      deploymentAuth: getDeploymentAuthLevel(project.deploymentAuth, isTeam),
-    },
-  });
-
-  const onSubmit: SubmitHandler<Inputs> = async (data) => {
-    const domain = `${data.domain}.${INTERNAL_DOMAIN_SUFFIX}`;
-    const savedDeploymentAuth = data.deploymentAuthEnabled
-      ? data.deploymentAuth
-      : DeploymentAuth.Public;
-
-    await Promise.all([
-      client.mutate({
-        mutation: UpdateProjectDomainMutation,
-        variables: {
-          input: {
-            projectId: project.id,
-            domain,
-          },
-        },
-      }),
-      client.mutate({
-        mutation: UpdateProjectMutation,
-        variables: {
-          projectId: project.id,
-          deploymentProductionBranchGlob:
-            data.noCustomDeploymentProductionBranchGlob
-              ? null
-              : data.deploymentProductionBranchGlob,
-          deploymentAuth: savedDeploymentAuth,
-        },
-      }),
-    ]);
-
-    form.reset(data);
-  };
-
-  const noCustomDeploymentProductionBranchGlob = form.watch(
-    "noCustomDeploymentProductionBranchGlob",
-  );
-  const deploymentAuthEnabled = form.watch("deploymentAuthEnabled");
-  const deploymentAuth = form.watch("deploymentAuth");
-  const deploymentAuthDef = DEPLOYMENT_AUTH_DEFS[deploymentAuth];
-  const deploymentProductionBranchGlobFieldProps = form.register(
-    "deploymentProductionBranchGlob",
-    {
-      required: { message: "Pattern required", value: true },
-    },
-  );
-
-  return (
-    <>
-      <Card>
-        <Form form={form} onSubmit={onSubmit}>
-          <CardBody>
-            <CardTitle>Deployments</CardTitle>
-            <CardParagraph>
-              Choose how Argos identifies and serves production deployments.
-            </CardParagraph>
-            <div className="rounded-sm border">
-              <section className="p-4">
-                <h3 className="mb-1 font-semibold">
-                  Deployment authentication
-                </h3>
-                <div className="flex flex-col gap-10 sm:flex-row sm:items-center">
-                  <FormSwitch
-                    control={form.control}
-                    name="deploymentAuthEnabled"
-                    label="Log in protection"
-                  />
-                  {deploymentAuthEnabled && (
-                    <SelectField
-                      control={form.control}
-                      name="deploymentAuth"
-                      aria-label="Deployment authentication level"
-                      className="sm:w-80"
-                    >
-                      <SelectButton className="w-full">
-                        {deploymentAuthDef.label}
-                      </SelectButton>
-                      <FieldError />
-                      <Popover className="max-w-sm">
-                        <ListBox>
-                          <ListBoxItem
-                            id={DeploymentAuth.DomainPrivate}
-                            textValue="Standard protection"
-                          >
-                            <ListBoxItemLabel>
-                              Standard protection
-                            </ListBoxItemLabel>
-                            <ListBoxItemDescription>
-                              Protect all except production custom domains for
-                              your project.
-                            </ListBoxItemDescription>
-                          </ListBoxItem>
-                          <ListBoxItem
-                            id={DeploymentAuth.Private}
-                            textValue="All deployments"
-                            isDisabled={!isTeam}
-                          >
-                            <ListBoxItemLabel>All deployments</ListBoxItemLabel>
-                            <ListBoxItemDescription>
-                              {isTeam
-                                ? "Protect all domains."
-                                : "Requires a team."}
-                            </ListBoxItemDescription>
-                          </ListBoxItem>
-                        </ListBox>
-                      </Popover>
-                    </SelectField>
-                  )}
-                </div>
-              </section>
-              <section className="border-t p-4">
-                <h3 className="mb-1 font-semibold">Production domain</h3>
-                <p className="text-low mb-4 text-sm">
-                  Internal domain used to serve your production deployment.
-                </p>
-                <FormTextInput
-                  control={form.control}
-                  {...form.register("domain", {
-                    required: "Please enter a domain slug",
-                    maxLength: {
-                      value: 48,
-                      message: "Domain slugs must be 48 characters or less",
-                    },
-                    pattern: {
-                      value: SLUG_REGEX,
-                      message:
-                        "Domain slugs must be lowercase, start and end with an alphanumeric character, and may contain dashes in the middle.",
-                    },
-                  })}
-                  label="Production domain"
-                  hiddenLabel
-                  addon={`.${INTERNAL_DOMAIN_SUFFIX}`}
-                  className="max-w-md"
-                />
-              </section>
-              <section className="border-t p-4">
-                <h3 className="mb-1 font-semibold">
-                  Production deployment branch
-                </h3>
-                <p className="text-low text-sm">
-                  Any deployment from a branch that matches the specified
-                  pattern will be treated as a production deployment.
-                </p>
-                <div className="mt-4">
-                  <FormSwitch
-                    control={form.control}
-                    name="noCustomDeploymentProductionBranchGlob"
-                    label={
-                      project.repository ? (
-                        <>
-                          Use{" "}
-                          {getRepositoryLabel(project.repository.__typename)}{" "}
-                          repository's default branch:{" "}
-                          <Code>{defaultDeploymentProductionBranchGlob}</Code>
-                        </>
-                      ) : (
-                        <>
-                          Use <Code>main</Code> as production deployment branch
-                        </>
-                      )
-                    }
-                  />
-                  {!noCustomDeploymentProductionBranchGlob && (
-                    <>
-                      <FormTextInput
-                        control={form.control}
-                        {...deploymentProductionBranchGlobFieldProps}
-                        ref={(element) => {
-                          deploymentProductionBranchGlobFieldProps.ref(element);
-                          if (element) {
-                            if (
-                              !noCustomDeploymentProductionBranchGlob &&
-                              form.formState.defaultValues
-                                ?.noCustomDeploymentProductionBranchGlob !==
-                                noCustomDeploymentProductionBranchGlob
-                            ) {
-                              element.focus();
-                            }
-                          }
-                        }}
-                        label="Production deployment branch pattern"
-                        className="mt-4"
-                      />
-                      <p className="text-low mt-2 text-sm">
-                        Use patterns like <Code>main</Code>,{" "}
-                        <Code>{`{main,production}`}</Code>, or{" "}
-                        <Code>release/**</Code>.
-                      </p>
-                    </>
-                  )}
-                </div>
-              </section>
-            </div>
-          </CardBody>
-          <FormCardFooter
-            control={form.control}
-            actions={<DisableProjectDeploymentsButton projectId={project.id} />}
-          />
-        </Form>
-      </Card>
-    </>
-  );
-}
-
-const DEPLOYMENT_AUTH_DEFS: Record<DeploymentAuthLevel, { label: string }> = {
-  [DeploymentAuth.DomainPrivate]: { label: "Standard protection" },
-  [DeploymentAuth.Private]: { label: "All deployments" },
-};
-
-function getDeploymentAuthLevel(
-  deploymentAuth: DeploymentAuth,
-  isTeam: boolean,
-): DeploymentAuthLevel {
-  return isTeam && deploymentAuth === DeploymentAuth.Private
-    ? DeploymentAuth.Private
-    : DeploymentAuth.DomainPrivate;
-}
-
-function getDomainSlug(domain: string | null | undefined) {
-  if (!domain) {
-    return "";
-  }
-
-  const suffix = `.${INTERNAL_DOMAIN_SUFFIX}`;
-  if (!domain.endsWith(suffix)) {
-    return domain;
-  }
-
-  return domain.slice(0, -suffix.length);
-}
-
-function ProjectDeploymentsDisabled(props: {
-  project: DocumentType<typeof _ProjectFragment>;
-}) {
-  const { project } = props;
-
-  return (
-    <Card>
-      <CardBody>
-        <CardTitle>Deployments</CardTitle>
-        <CardParagraph>
-          Enable deployments on this project to deploy static environments like
-          Storybook.
-        </CardParagraph>
-      </CardBody>
-      <CardFooter className="flex items-center justify-between gap-4">
-        <div />
-        <div className="flex items-center justify-end gap-4">
-          <EnableProjectDeploymentsButton projectId={project.id} />
-        </div>
-      </CardFooter>
-    </Card>
-  );
-}
-
-function DisableProjectDeploymentsConfirmButton(props: { projectId: string }) {
-  const { projectId } = props;
-  const client = useApolloClient();
-  const state = useOverlayTriggerState();
-
-  return (
-    <Button
-      variant="destructive"
-      onAction={async () => {
-        await client.mutate({
-          mutation: UpdateProjectMutation,
-          variables: {
-            projectId,
-            deploymentEnabled: false,
-          },
-        });
-        state.close();
-      }}
-    >
-      Disable
-    </Button>
-  );
-}
-
-function DisableProjectDeploymentsButton(props: { projectId: string }) {
-  const { projectId } = props;
-  return (
-    <DialogTrigger>
-      <Button variant="secondary">Disable</Button>
-      <Modal>
-        <Dialog size="medium">
-          <DialogBody>
-            <DialogTitle>Disable Deployments</DialogTitle>
-            <DialogText>
-              Deployment URLs for this project will become inaccessible after
-              deployments are disabled, although some existing links may
-              continue to work briefly while cached resolutions expire.
-            </DialogText>
-            <div className="bg-danger-hover text-danger-low my-4 rounded-sm p-2">
-              <strong>Warning:</strong> Existing deployment links may remain
-              accessible briefly until cached resolutions expire.
-            </div>
-          </DialogBody>
-          <DialogFooter>
-            <DialogDismiss>Cancel</DialogDismiss>
-            <DisableProjectDeploymentsConfirmButton projectId={projectId} />
-          </DialogFooter>
-        </Dialog>
-      </Modal>
-    </DialogTrigger>
-  );
-}
-
-function EnableProjectDeploymentsButton(props: { projectId: string }) {
-  const { projectId } = props;
-  const client = useApolloClient();
-
-  return (
-    <Button
-      onAction={async () => {
-        await client.mutate({
-          mutation: UpdateProjectMutation,
-          variables: {
-            projectId,
-            deploymentEnabled: true,
-          },
-        });
-      }}
-    >
-      Enable
-    </Button>
-  );
-}
